requirements.txt
flask==3.1.2
pdfplumber==0.10.1
python-docx==0.8.12
